api/requirements.txt
# requirements.txt

packaging==23.2
pillow==10.2.0
pyperclip==1.8.2
pytesseract==0.3.10


# requirements.txt

aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
easyocr==1.7.1
fastapi==0.109.2
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
Jinja2==3.1.3
keyboard==0.13.5
lazy_loader==0.3
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
openai==0.28.0
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
pyclipper==1.3.0.post5
pydantic==2.6.1
pydantic_core==2.16.2
pyperclip==1.8.2
pytesseract==0.3.10
python-bidi==0.4.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
sympy==1.12
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
yarl==1.9.4
